Moved IsEligibleForSurrogate to its own separate file
