+++ Source/ILAssist/Attributes.cs
@@ -45,28 +45,5 @@ namespace Surrogate.ILAssist
 
 		internal static List<T> FindAttributes<T>(this ParameterInfo Parameter)
 		=> Parameter.FindAttributes(typeof(T)).Cast<T>().ToList();
-
-		internal static bool IsEligibleForSurrogate(this MethodInfo Method)
-		{
-			var parameterAttributes = Method.GetParameters().SelectMany(i => i.FindAttributes<IParameterSurrogate>());
-			if (parameterAttributes.Count() > 0)
-				return true;
-
-			var methodAttributes = Method.FindAttributes<IMethodSurrogate>();
-			if (methodAttributes.Count > 0)
-				return true;
-
-			var returnAttributes = Method.FindAttributes<IReturnSurrogate>();
-			if (returnAttributes.Count > 0)
-				return true;
-
-			return false;
-		}
-
-		internal static bool IsEligibleForSurrogate(this Type BaseType)
-		{
-			var methods = BaseType.GetMethods(BindingFlags.Instance | BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.DeclaredOnly);
-			return methods.Select(i => i.IsEligibleForSurrogate()).Aggregate((current, next) => current || next);
-		}
 	}
 }
